+++ src/fs/coss/store_io_coss.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: store_io_coss.cc,v 1.12 2001/10/24 05:26:22 hno Exp $
+ * $Id: store_io_coss.cc,v 1.13 2001/10/24 07:45:37 hno Exp $
  *
  * DEBUG: section 81    Storage Manager COSS Interface
  * AUTHOR: Eric Stern
@@ -109,8 +109,8 @@ storeCossAllocate(SwapDir * SD, const StoreEntry * e, int which)
 	 */
 	cs->current_membuf->flags.full = 1;
 	cs->current_offset = cs->current_membuf->diskend + 1;
-	debug(81, 2) ("storeCossAllocate: New offset - %d\n",
-	    cs->current_offset);
+	debug(81, 2) ("storeCossAllocate: New offset - %ld\n",
+	    (long int) cs->current_offset);
 	newmb = storeCossCreateMemBuf(SD, cs->current_offset, checkf, &coll);
 	cs->current_membuf = newmb;
     }
@@ -159,7 +159,7 @@ storeCossCreate(SwapDir * SD, StoreEntry * e, STFNCB * file_callback, STIOCB * c
     sio->st_size = objectLen(e) + e->mem_obj->swap_hdr_sz;
     sio->swap_dirn = SD->index;
     sio->swap_filen = storeCossAllocate(SD, e, COSS_ALLOC_ALLOCATE);
-    debug(81, 3) ("storeCossCreate: offset %d, size %d, end %d\n", sio->swap_filen, sio->st_size, sio->swap_filen + sio->st_size);
+    debug(81, 3) ("storeCossCreate: offset %d, size %ld, end %ld\n", sio->swap_filen, (long int) sio->st_size, (long int) (sio->swap_filen + sio->st_size));
 
     sio->callback = callback;
     sio->file_callback = file_callback;
@@ -282,7 +282,7 @@ storeCossRead(SwapDir * SD, storeIOState * sio, char *buf, size_t size, off_t of
     assert(sio->read.callback_data == NULL);
     sio->read.callback = callback;
     sio->read.callback_data = callback_data;
-    debug(81, 3) ("storeCossRead: offset %ld\n", (long int)offset);
+    debug(81, 3) ("storeCossRead: offset %ld\n", (long int) offset);
     sio->offset = offset;
     cstate->flags.reading = 1;
     if ((offset + size) > sio->st_size)
@@ -322,7 +322,7 @@ storeCossWrite(SwapDir * SD, storeIOState * sio, char *buf, size_t size, off_t o
      */
     assert(sio->e->mem_obj->object_sz != -1);
 
-    debug(81, 3) ("storeCossWrite: offset %ld, len %lu\n", (long int)sio->offset, (unsigned long int)size);
+    debug(81, 3) ("storeCossWrite: offset %ld, len %lu\n", (long int) sio->offset, (unsigned long int) size);
     diskoffset = sio->swap_filen + sio->offset;
     dest = storeCossMemPointerFromDiskOffset(SD, diskoffset, &membuf);
     assert(dest != NULL);
@@ -476,8 +476,8 @@ static void
 storeCossWriteMemBuf(SwapDir * SD, CossMemBuf * t)
 {
     CossInfo *cs = (CossInfo *) SD->fsdata;
-    debug(81, 3) ("storeCossWriteMemBuf: offset %d, len %d\n",
-	t->diskstart, t->diskend - t->diskstart);
+    debug(81, 3) ("storeCossWriteMemBuf: offset %ld, len %ld\n",
+	(long int) t->diskstart, (long int) (t->diskend - t->diskstart));
     t->flags.writing = 1;
     /* Remember that diskstart/diskend are block offsets! */
     a_file_write(&cs->aq, cs->fd, t->diskstart, &t->buffer,
@@ -491,7 +491,7 @@ storeCossWriteMemBufDone(int fd, int errflag, size_t len, void *my_data)
     CossMemBuf *t = my_data;
     CossInfo *cs = (CossInfo *) t->SD->fsdata;
 
-    debug(81, 3) ("storeCossWriteMemBufDone: buf %p, len %ld\n", t, (long int)len);
+    debug(81, 3) ("storeCossWriteMemBufDone: buf %p, len %ld\n", t, (long int) len);
     if (errflag)
 	debug(81, 0) ("storeCossMemBufWriteDone: got failure (%d)\n", errflag);
 
@@ -512,7 +512,7 @@ storeCossCreateMemBuf(SwapDir * SD, size_t start,
     CBDATA_INIT_TYPE_FREECB(CossMemBuf, NULL);
     newmb = cbdataAlloc(CossMemBuf);
     newmb->diskstart = start;
-    debug(81, 3) ("storeCossCreateMemBuf: creating new membuf at %d\n", newmb->diskstart);
+    debug(81, 3) ("storeCossCreateMemBuf: creating new membuf at %ld\n", (long int) newmb->diskstart);
     debug(81, 3) ("storeCossCreateMemBuf: at %p\n", newmb);
     newmb->diskend = newmb->diskstart + COSS_MEMBUF_SZ - 1;
     newmb->flags.full = 0;
@@ -525,7 +525,7 @@ storeCossCreateMemBuf(SwapDir * SD, size_t start,
     /* Print out the list of membufs */
     for (m = cs->membufs.head; m; m = m->next) {
 	t = m->data;
-	debug(81, 3) ("storeCossCreateMemBuf: membuflist %d lockcount %d\n", t->diskstart, t->lockcount);
+	debug(81, 3) ("storeCossCreateMemBuf: membuflist %ld lockcount %d\n", (long int) t->diskstart, t->lockcount);
     }
 
     /*
